Extract FormProfile to its own module
* Add `valuesArePresent` to move form validation into `Model.FormProfile`
