+++ croniker.cabal
@@ -34,6 +34,7 @@ library
                      Import
                      Import.NoFoundation
                      Model
+                     Model.FormProfile
                      Model.Profile
                      Model.User
                      Settings
